Switch to 24HR time
Seems more intuitive for this particular application.
Dealing with AM/PM would either require explicit
definition at the command prompt (which is more typing),
or to guess based on current time (which can feel
unpredictable).

It's a possibility to guess based on current time, and
allow optional explicit definitions at the command line.
Can be considered in the future.